implement vote_message signature
